# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SteamShare API - Steam account marketplace with points, XP, and badges
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /accounts:
    get:
      operationId: listAccounts
      tags: [accounts]
      summary: List steam accounts (recent or filtered by game)
      parameters:
        - name: game
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [recent, popular, free, points]
            default: recent
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of accounts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountListResponse"

    post:
      operationId: createAccount
      tags: [accounts]
      summary: Post a steam account listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /accounts/{accountId}:
    get:
      operationId: getAccount
      tags: [accounts]
      summary: Get a single steam account listing
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

    delete:
      operationId: deleteAccount
      tags: [accounts]
      summary: Delete a steam account listing (owner or admin)
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /accounts/{accountId}/claim:
    post:
      operationId: claimAccount
      tags: [accounts]
      summary: Claim/buy a steam account using points
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Claimed successfully — returns steam credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClaimResult"
        "400":
          description: Not enough points or already claimed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /accounts/{accountId}/like:
    post:
      operationId: likeAccount
      tags: [accounts]
      summary: Like a steam account listing
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Liked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"

    delete:
      operationId: unlikeAccount
      tags: [accounts]
      summary: Remove like from a steam account listing
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Like removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"

  # Comments
...
  /users/{userId}/accounts:
    get:
      operationId: getUserAccounts
      tags: [users]
      summary: Get all accounts posted by a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"

  # Ad Links
...
  /accounts/verify-credentials:
    post:
      operationId: verifyCredentials
      tags: [accounts]
      summary: Check if Steam credentials are valid before posting
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyCredentialsInput"
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyCredentialsResult"

  # Giveaways
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessMessage:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Account:
      type: object
      required: [id, userId, username, title, description, games, pointsCost, isAvailable, likesCount, claimsCount, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        username:
          type: string
        posterUsername:
          type: string
        posterAvatarUrl:
          type: ["string", "null"]
        title:
          type: string
        description:
          type: string
        games:
          type: array
          items:
            type: string
        pointsCost:
          type: integer
        isAvailable:
          type: boolean
        likesCount:
          type: integer
        claimsCount:
          type: integer
        viewCount:
          type: integer
        workingVotes:
          type: integer
        notWorkingVotes:
          type: integer
        unlockMethod:
          type: string
          enum: [login, like, comment]
        userHasLiked:
          type: boolean
        userHasCommented:
          type: boolean
        createdAt:
          type: string
          format: date-time
    AccountListResponse:
      type: object
      required: [accounts, total, page, limit]
      properties:
        accounts:
          type: array
          items:
            $ref: "#/components/schemas/Account"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    AccountInput:
      type: object
      required: [title, description, games, pointsCost, steamUsername, steamPassword]
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 100
        description:
          type: string
          maxLength: 1000
        games:
          type: array
          items:
            type: string
          minItems: 1
        pointsCost:
          type: integer
          minimum: 0
        steamUsername:
          type: string
        steamPassword:
          type: string
        unlockMethod:
          type: string
          enum: [login, like, comment]
    ClaimResult:
      type: object
      required: [steamUsername, steamPassword, pointsSpent]
      properties:
        steamUsername:
          type: string
        steamPassword:
          type: string
        pointsSpent:
          type: integer
...
    VerifyCredentialsInput:
      type: object
      required: [steamUsername, steamPassword]
      properties:
        steamUsername:
          type: string
        steamPassword:
          type: string
    VerifyCredentialsResult:
      type: object
      required: [status, message]
      properties:
        status:
          type: string
          enum: [valid, invalid, rate_limited, error]
        message:
          type: string
        games:
          type: array
          items:
            type: string
        steamid:
          type: string
        isFamilyShare:
          type: boolean
...